requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
environs==9.5.0
inflection==0.5.1
jsonschema==4.17.3
marshmallow==3.19.0
packaging==23.1
Pillow==9.5.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
